Changes
This added the dungeon lvl & hell lvl finished with enemies for each level.
Snake for dungeon and Demon for hell. 
There's the dungeonKeeper & hellKeeper for the player to advance to next levels with a guide.
Allows player to advance to next level once they collected enough pts and want to advance.